Update elastic_server_deb.rst
Tried to be more consistent with for later wording "For more info please see".
In line 30 we are asked to "install curl apt-transport-https", but we have already done this on line 16 of "https://documentation.wazuh.com/current/installation-guide/installing-wazuh-server/wazuh_server_deb.html". 
Thanks again for all your hard work.