Update Puppet Documentation
If we followed Puppet's documentation step by step we couldn't install either Wazuh's manager or Wazuh's agent.

We have updated the links to install version 5 of Puppet, as well as advise for any operating system to use this version of Puppet or a greater version.

We have added necessary commands for the correct execution of the installation.

We have added tips and warnings regarding outdated components, such as the installation of Wazuh 2.1.1.

We have added a line with which to complete the manifests of the repository for Wazuh manager installations on Debian and Ubuntu systems.