General corrections
- Confusing and wrong list order
It has been fixed the list order bug. The problem was in the CSS code

- Fix code block when it's inside something (twice in this page)
It no longer appears the weird padding in the code blocks

- Weird padding
It has been fixed the different padding in some pages

- Replace ul for ol
In some lists appeared numbers and points together. It is already
corrected

- Some pages not showing the section on the toctree (sidebar menu)
Page by page it has been verified that all show the tree correctly, as
in the old documentation

- Fix focus when clicking on links (next and prev)
It has been removed the default style of the focus and it has been added
an underline style

- Replace naked link for Transforming Data at the end of
/installation-guide/installing-elastic-stack/transform_logstash.html
The naked link now it is dressed :)

- Fix list inside a note
When a list is on a note now has more padding in the left

- I have reviewed page by page to check all these bugs

- Also, I have fixed other minor bugs of CSS