Update wazuh_server_rpm_suse_opensuse.rst
Hi, Team. Over this branch (not updated to web server yet) I realized the repository path at disabling Wazuh repo is not correct because we are working with OpenSUSE (zypper) and not with CentOS/Fedora (yum). Same with the creation of Elastic repo directory. I also added the way to disable the Elastic Stack repo (optional step 8). 
Greetings.