Added how to create custom scripts
Hi Team.

I considered adding how to create a custom active-response script would be quite helpful. I added the parameters sent to an agent node to execute this kind of scripts. The way to give that file permissions and the procedure are also explained below. This will gather the information contained in FAQ for higher visibility.

Best regards,
Pablo Rodríguez