Update what_is_it.rst
I found it confusing that SCA results are *not* shown in the Wazuh App's "overview" screen (as are many other types of results), but *are* shown in individual agent overview screens.  I don't know whether or not this is intentional, but if it is (or even if it isn't, until a correction is made), I strongly recommend documenting this fact here and elsewhere.