Update how_it_works.rst
Hello team,

Strongly recommends that you write clearly when it comes to a single manager or a cluster. Whenever you write, pay attention to it as it causes some confusion. Especially that the CVE module works on workers and managers, so if the manager drops out for some reason, the module still works, but on workers.