Update securing-api.rst
The current securing-api documentation states that the Wazuh API will be initialized with a default password of "wazuh-wiu". This is no longer the case, as Wazuh now creates a unique enrollment password for the API user. I have proposed a change to this page to reflect this change, which will prevent users not being able to log-into the API, as well as give them information on where to find their enrollment password so that the can use the API.